+++ ignite/pkg/cosmosclient/bank_test.go
@@ -5,9 +5,9 @@ import (
 	"testing"
 
 	"cosmossdk.io/math"
+	banktypes "cosmossdk.io/x/bank/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/types/query"
-	banktypes "github.com/cosmos/cosmos-sdk/x/bank/types"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )