+++ ignite/pkg/cosmosgen/template.go
@@ -66,7 +66,7 @@ func (t templateWriter) Write(destDir, protoPath string, data interface{}) error
 
 			return strcase.ToCamel(replacer.Replace(word))
 		},
-		"resolveFile": func(fullPath string) string {
+		"resolveFile": func(fullPath string) string { // TODO to check
 			rel, _ := filepath.Rel(protoPath, fullPath)
 			rel = strings.TrimSuffix(rel, ".proto")
 			return rel