+++ integration/ibc/cmd_ibc_test.go
@@ -10,6 +10,8 @@ import (
 )
 
 func TestCreateModuleWithIBC(t *testing.T) {
+	t.Skip("skipping test as IBC isn't available with v0.52 yet") // https://github.com/ignite/cli/pull/4289
+
 	var (
 		env = envtest.New(t)
 		app = env.Scaffold("github.com/test/blogibc")
@@ -107,6 +109,8 @@ func TestCreateModuleWithIBC(t *testing.T) {
 }
 
 func TestCreateIBCPacket(t *testing.T) {
+	t.Skip("skipping test as IBC isn't available with v0.52 yet") // https://github.com/ignite/cli/pull/4289
+
 	var (
 		env = envtest.New(t)
 		app = env.Scaffold("github.com/test/blogibcb")